requirements.txt
streamlit>=1.32
opencv-python-headless>=4.9
scikit-image>=0.22
scikit-learn>=1.4
numpy>=1.26
Pillow>=10.0
joblib>=1.3
matplotlib>=3.8
seaborn>=0.13
